fix: update in mem guardrail params
