+++ litellm/proxy/guardrails/init_guardrails.py
@@ -89,10 +89,12 @@ def init_guardrails_v2(
     all_guardrails: List[Dict],
     config_file_path: Optional[str] = None,
 ):
+    from litellm.proxy.guardrails.guardrail_registry import IN_MEMORY_GUARDRAIL_HANDLER
+
     guardrail_list: List[Guardrail] = []
 
     for guardrail in all_guardrails:
-        initialized_guardrail = InitializeGuardrails.initialize_guardrail(
+        initialized_guardrail = IN_MEMORY_GUARDRAIL_HANDLER.initialize_guardrail(
             guardrail=guardrail,
             config_file_path=config_file_path,
         )
@@ -100,118 +102,3 @@ def init_guardrails_v2(
             guardrail_list.append(initialized_guardrail)
 
     verbose_proxy_logger.info(f"\nGuardrail List:{guardrail_list}\n")
-
-
-class InitializeGuardrails:
-    """
-    Class that handles initializing guardrails and adding them to the CallbackManager
-    """
-
-    @staticmethod
-    def initialize_guardrail(
-        guardrail: Dict,
-        config_file_path: Optional[str] = None,
-    ) -> Optional[Guardrail]:
-        """
-        Initialize a guardrail from a dictionary and add it to the litellm callback manager
-
-        Returns a Guardrail object if the guardrail is initialized successfully
-        """
-        litellm_params_data = guardrail["litellm_params"]
-        verbose_proxy_logger.debug("litellm_params= %s", litellm_params_data)
-
-        litellm_params = LitellmParams(**litellm_params_data)
-
-        if (
-            "category_thresholds" in litellm_params_data
-            and litellm_params_data["category_thresholds"]
-        ):
-            lakera_category_thresholds = LakeraCategoryThresholds(
-                **litellm_params_data["category_thresholds"]
-            )
-            litellm_params.category_thresholds = lakera_category_thresholds
-
-        if litellm_params.api_key and litellm_params.api_key.startswith("os.environ/"):
-            litellm_params.api_key = str(get_secret(litellm_params.api_key))
-
-        if litellm_params.api_base and litellm_params.api_base.startswith(
-            "os.environ/"
-        ):
-            litellm_params.api_base = str(get_secret(litellm_params.api_base))
-
-        guardrail_type = litellm_params.guardrail
-        if guardrail_type is None:
-            raise ValueError("guardrail_type is required")
-
-        initializer = guardrail_initializer_registry.get(guardrail_type)
-
-        if initializer:
-            initializer(litellm_params, guardrail)
-        elif isinstance(guardrail_type, str) and "." in guardrail_type:
-            InitializeGuardrails.initialize_custom_guardrail(
-                guardrail=guardrail,
-                guardrail_type=guardrail_type,
-                litellm_params=litellm_params,
-                config_file_path=config_file_path,
-            )
-        else:
-            raise ValueError(f"Unsupported guardrail: {guardrail_type}")
-
-        parsed_guardrail = Guardrail(
-            guardrail_name=guardrail["guardrail_name"],
-            litellm_params=litellm_params,
-        )
-
-        return parsed_guardrail
-
-    @staticmethod
-    def initialize_custom_guardrail(
-        guardrail: Dict,
-        guardrail_type: str,
-        litellm_params: LitellmParams,
-        config_file_path: Optional[str] = None,
-    ) -> None:
-        """
-        Initialize a Custom Guardrail from a python file
-
-        This initializes it by adding it to the litellm callback manager
-        """
-        if not config_file_path:
-            raise Exception(
-                "GuardrailsAIException - Please pass the config_file_path to initialize_guardrails_v2"
-            )
-
-        _file_name, _class_name = guardrail_type.split(".")
-        verbose_proxy_logger.debug(
-            "Initializing custom guardrail: %s, file_name: %s, class_name: %s",
-            guardrail_type,
-            _file_name,
-            _class_name,
-        )
-
-        directory = os.path.dirname(config_file_path)
-        module_file_path = os.path.join(directory, _file_name) + ".py"
-
-        spec = importlib.util.spec_from_file_location(_class_name, module_file_path)  # type: ignore
-        if not spec:
-            raise ImportError(
-                f"Could not find a module specification for {module_file_path}"
-            )
-
-        module = importlib.util.module_from_spec(spec)  # type: ignore
-        spec.loader.exec_module(module)  # type: ignore
-        _guardrail_class = getattr(module, _class_name)
-
-        mode = litellm_params.mode
-        if mode is None:
-            raise ValueError(
-                f"mode is required for guardrail {guardrail_type} please set mode to one of the following: {', '.join(GuardrailEventHooks)}"
-            )
-
-        default_on = litellm_params.default_on
-        _guardrail_callback = _guardrail_class(
-            guardrail_name=guardrail["guardrail_name"],
-            event_hook=mode,
-            default_on=default_on,
-        )
-        litellm.logging_callback_manager.add_litellm_callback(_guardrail_callback)  # type: ignore
